fixed problems with DAO tests
*added test start logging

+++ dao/src/main/java/com/dao/project/ProjectDAOImpl.java
@@ -8,9 +8,11 @@ import org.hibernate.query.Query;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
+import javax.transaction.Transactional;
 import java.util.List;
 
 @Repository
+@Transactional
 public class ProjectDAOImpl implements ProjectDAO {
     private SessionFactory sessionFactory;
 
@@ -86,11 +88,15 @@ public class ProjectDAOImpl implements ProjectDAO {
     @Override
     public boolean remove(Project project) {
         Session session = sessionFactory.getCurrentSession();
-        try {
-            session.delete(project);
-            return true;
-        } catch (Exception e) {
-            return false;
+        project = (Project) session.merge(project);
+        if(project == null) return false;
+        else {
+            try {
+                session.delete(project);
+                return true;
+            } catch (Exception e) {
+                return false;
+            }
         }
     }
 
@@ -102,6 +108,7 @@ public class ProjectDAOImpl implements ProjectDAO {
         else if (project.isActive()) {
             try {
                 project.setActive(false);
+                session.saveOrUpdate(project);
             } catch (Exception e) {
                 return false;
             }
@@ -112,12 +119,13 @@ public class ProjectDAOImpl implements ProjectDAO {
     @Override
     public boolean inactivate(Project project) {
         Session session = sessionFactory.getCurrentSession();
-        Project projectFromDB = session.get(Project.class, project.getId());
-        if (!project.isActive()) return true;
-        else if (projectFromDB == null) return false;
+        project = (Project) session.merge(project);
+        if (project == null) return false;
+        else if (!project.isActive()) return true;
         else {
             try {
                 project.setActive(false);
+                session.saveOrUpdate(project);
             } catch (Exception e) {
                 return false;
             }
@@ -133,6 +141,7 @@ public class ProjectDAOImpl implements ProjectDAO {
         else if (!project.isActive()) {
             try {
                 project.setActive(true);
+                session.saveOrUpdate(project);
             } catch (Exception e) {
                 return false;
             }
@@ -143,12 +152,13 @@ public class ProjectDAOImpl implements ProjectDAO {
     @Override
     public boolean activate(Project project) {
         Session session = sessionFactory.getCurrentSession();
-        Project projectFromDB = session.get(Project.class, project.getId());
-        if (project.isActive()) return true;
-        else if (projectFromDB == null) return false;
+        project = (Project) session.merge(project);
+        if (project == null) return false;
+        else if (project.isActive()) return true;
         else {
             try {
                 project.setActive(true);
+                session.saveOrUpdate(project);
             } catch (Exception e) {
                 return false;
             }
